% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvis-export.R
\name{cuvis_export_envi}
\alias{cuvis_export_envi}
\title{Export a Measurement to ENVI Format}
\usage{
cuvis_export_envi(measurement, dir, permissive = TRUE)
}
\arguments{
\item{measurement}{A cuvis_measurement object.}

\item{dir}{Character. Output directory.}

\item{permissive}{Logical. Allow export even if some data is missing.
Default \code{TRUE}.}
}
\value{
Invisible \code{dir}.
}
\description{
Writes the measurement data as an ENVI header/binary file pair.
}
