% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvis-export.R
\name{cuvis_export_tiff}
\alias{cuvis_export_tiff}
\title{Export a Measurement to TIFF Format}
\usage{
cuvis_export_tiff(
  measurement,
  dir,
  format = c("MultiChannel", "Single", "MultiPage"),
  compression = c("None", "LZW"),
  permissive = TRUE
)
}
\arguments{
\item{measurement}{A cuvis_measurement object.}

\item{dir}{Character. Output directory.}

\item{format}{Character. TIFF format: \code{"MultiChannel"} (default),
\code{"Single"}, or \code{"MultiPage"}.}

\item{compression}{Character. Compression mode: \code{"None"} (default)
or \code{"LZW"}.}

\item{permissive}{Logical. Default \code{TRUE}.}
}
\value{
Invisible \code{dir}.
}
\description{
Export a Measurement to TIFF Format
}
