% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvis-processing.R
\name{cuvis_has_reference}
\alias{cuvis_has_reference}
\title{Check if a Reference is Set}
\usage{
cuvis_has_reference(
  ctx,
  type = c("dark", "white", "white_dark", "sprad", "distance")
)
}
\arguments{
\item{ctx}{A cuvis_processing_context object.}

\item{type}{Character. Reference type to check.}
}
\value{
Logical.
}
\description{
Check if a Reference is Set
}
