% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvis-init.R
\name{cuvis_is_initialized}
\alias{cuvis_is_initialized}
\title{Is the CUVIS SDK Currently Initialised?}
\usage{
cuvis_is_initialized()
}
\value{
Logical.
}
\description{
Is the CUVIS SDK Currently Initialised?
}
\examples{
cuvis_is_initialized()
}
