% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvis-enums.R
\name{cuvis_session_item_types}
\alias{cuvis_session_item_types}
\title{CUVIS Session Item Types}
\usage{
cuvis_session_item_types
}
\description{
Named integer vector used to filter session file frames.
}
\keyword{internal}
