% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvis-init.R
\name{cuvis_version}
\alias{cuvis_version}
\title{Get CUVIS SDK Version}
\usage{
cuvis_version()
}
\value{
Character string with the SDK version.
}
\description{
Note: The SDK must be initialized with \code{\link[=cuvis_init]{cuvis_init()}} before calling
this function.
}
