% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuvis.r-package.R
\docType{package}
\name{cuvis.r-package}
\alias{cuvis.r}
\alias{cuvis.r-package}
\title{cuvis.r: R Bindings for the Cubert CUVIS Hyperspectral SDK}
\description{
Provides R bindings to the Cubert CUVIS C SDK for reading, processing, and exporting hyperspectral imaging data from Cubert snapshot cameras. Supports loading proprietary '.cu3s' session files, applying the full calibration pipeline (dark subtraction, white reference normalization, spectral radiance, reflectance), and exporting to standard formats (ENVI, TIFF). Requires the Cubert CUVIS SDK to be installed separately. This package mirrors the API of 'cuvis.python' (\url{https://github.com/cubert-hyperspectral/cuvis.python}) for R users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-heller/cuvis.r}
  \item \url{https://r-heller.github.io/cuvis.r/}
  \item Report bugs at \url{https://github.com/r-heller/cuvis.r/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

}
\keyword{internal}
